feat: add FlickerAPI
