+++ package.json
@@ -9,6 +9,7 @@
     "axios": "^1.5.0",
     "firebase": "^10.8.1",
     "firebase-tools": "^13.4.1",
+    "flickr-sdk": "^7.0.0-beta.7",
     "react": "^18.2.0",
     "react-anchor-navigation": "^0.2.7",
     "react-dom": "^18.2.0",
